+++ src/admin_websocket.rs
@@ -3,7 +3,8 @@ use anyhow::Result;
 use holo_hash::DnaHash;
 use holochain_conductor_api::{
     AdminRequest, AdminResponse, AppAuthenticationTokenIssued, AppInfo, AppInterfaceInfo,
-    AppStatusFilter, CompatibleCells, IssueAppAuthenticationTokenPayload, StorageInfo,
+    AppStatusFilter, CompatibleCells, IssueAppAuthenticationTokenPayload, RevokeAgentKeyPayload,
+    StorageInfo,
 };
 use holochain_types::websocket::AllowedOrigins;
 use holochain_types::{
@@ -97,6 +98,22 @@ impl AdminWebsocket {
         let response = self.send(AdminRequest::GenerateAgentPubKey).await?;
         match response {
             AdminResponse::AgentPubKeyGenerated(key) => Ok(key),
+            _ => unreachable!("Unexpected response {:?}", response),
+        }
+    }
+
+    pub async fn revoke_agent_key(
+        &self,
+        app_id: String,
+        agent_key: AgentPubKey,
+    ) -> ConductorApiResult<Vec<(CellId, String)>> {
+        let response = self
+            .send(AdminRequest::RevokeAgentKey(Box::new(
+                RevokeAgentKeyPayload { app_id, agent_key },
+            )))
+            .await?;
+        match response {
+            AdminResponse::AgentKeyRevoked(errors) => Ok(errors),
             _ => unreachable!("Unexpected response {:?}", response),
         }
     }